pyproject.toml
[tool.poetry]
name = "analyst"
version = "0.1.0"
description = ""
authors = ["Laurent Arthur <laurent.arthur75@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
cachetools = "^4.2.2"
pandas = "^1.2.4"
requests = "^2.25.1"
numpy = "^1.20.3"
tabulate = "^0.8.9"
hartware-lib = "0.3.1"
matplotlib = "^3.5.1"
redis = "^4.2.2"
#dash = "^2.5.0"
#dash-bootstrap-components = "^1.1.0"
#PySide6 = "^6.3.0"
#pyqtgraph = "^0.12.4"
jupyterlab = "^3.4.4"
pydantic = "^1.9.0"
jupyterthemes = "^0.20.0"
websockets = "^10.3"
seaborn = "^0.11.2"
click = "8.0.4"
aiohttp = "^3.8.1"
pytest-asyncio = "^0.19.0"
pika = "^1.3.0"
prompt-toolkit = "^3.0.30"
aio-pika = "^8.1.1"
py-flags = "^1.1.4"
motor = "^3.0.0"
PyJWT = "^2.4.0"
freezegun = "^1.2.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
isort = "^5.9.1"
ipython = "^7.24.1"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
mypy = "^0.910"
black = "^21.7b0"
piprot = "^0.9.11"
types-cachetools = "^4.2.9"
types-requests = "^2.27.9"
pytest-env = "^0.6.2"
types-redis = "^4.3.13"
types-tabulate = "^0.8.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
